refactor: move mock responses to seperate file
